refactor: rename DatabaseCommandService to kebab-case naming convention
