+++ commands/cmd/cmd-operations.js
@@ -8,7 +8,7 @@ import { createToggleMenu } from '../../utils/toggle-menu.js'
 import { createSelectionTitle } from '../../utils/menu-helpers.js'
 import { outputHandler } from '../../core/output-handler.js'
 import { color } from '../../config/color.js'
-import { databaseCommandService } from '../../services/DatabaseCommandService.js'
+import { databaseCommandService } from '../../services/database-command-service.js'
 import { createSpinner } from '../../utils/spinner.js'
 import { APP_CONSTANTS } from '../../config/constants.js'
 import { getAllSystemCommandNames } from '../../utils/system-commands.js'
